export component KycRequirement
